# postman/openapi/skolo_1_0_3.openapi.yaml
openapi: 3.0.3
info:
  title: Skolo API
  version: 1.0.3
  description: School transport tracking and management API
  contact:
    name: Skolo Team
servers:
  -
    url: "http://localhost:3000/api"
    description: Development
  -
    url: "https://api.skolo.com/api"
    description: Production
...
paths:
...
  /admin/school-assignments/{schoolId}/driver/{driverId}:
    get:
      tags:
        - Admin
      summary: Get Driver
      description: "GET /school-assignments/:schoolId/driver/:driverId"
      operationId: get_admin_school-assignments__schoolId_driver__driverId
      parameters:
        -
          name: schoolId
          in: path
          required: true
          schema:
            type: string
          description: SchoolId identifier
        -
          name: driverId
          in: path
          required: true
          schema:
            type: string
          description: DriverId identifier
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        500:
          description: Internal server error
...
  /admin/school-drivers/{schoolId}:
    get:
      tags:
        - Admin
      summary: Get School Drivers
      description: "GET /school-drivers/:schoolId"
      operationId: get_admin_school-drivers__schoolId
      parameters:
        -
          name: schoolId
          in: path
          required: true
          schema:
            type: string
          description: SchoolId identifier
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        500:
          description: Internal server error
...
  /admin/school-subscriptions/school/{schoolId}:
    get:
      tags:
        - Admin
      summary: Get School
      description: "GET /school-subscriptions/school/:schoolId"
      operationId: get_admin_school-subscriptions_school__schoolId
      parameters:
        -
          name: schoolId
          in: path
          required: true
          schema:
            type: string
          description: SchoolId identifier
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        500:
          description: Internal server error
...
  /admin/schools/{school_id}:
    get:
      tags:
        - Admin
      summary: Get Schools
      description: "GET /schools/:school_id"
      operationId: get_admin_schools__school_id
      parameters:
        -
          name: school_id
          in: path
          required: true
          schema:
            type: string
          description: School Id identifier
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        500:
          description: Internal server error
    put:
      tags:
        - Admin
      summary: Update Schools
      description: "\n\n📋 Request Body Fields:\n• school_name: string (optional)\n• address: string (optional)\n• city: string (optional)\n• state: string (optional)\n• latitude: number (optional)\n• longitude: number (optional)\n• contact_number: string (optional)\n• email: string (optional)"
      operationId: put_admin_schools__school_id
      parameters:
        -
          name: school_id
          in: path
          required: true
          schema:
            type: string
          description: School Id identifier
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        500:
          description: Internal server error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                school_name:
                  type: string
                  minimum: 3
                  maximum: 200
                address:
                  type: string
                  maximum: 255
                city:
                  type: string
                  maximum: 100
                state:
                  type: string
                  maximum: 100
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
                contact_number:
                  type: string
                  pattern: ^[+]?[0-9]{10,15}$
                email:
                  type: string
                  format: email
            example:
              school_name: Delhi Public School
              address: 123 Main Street, Mumbai
              city: Mumbai
              state: Maharashtra
              latitude: 19.076
              longitude: 72.8777
              contact_number: +919876543210
              email: user@example.com
    delete:
      tags:
        - Admin
      summary: Delete Schools
      description: "DELETE /schools/:school_id"
      operationId: delete_admin_schools__school_id
      parameters:
        -
          name: school_id
          in: path
          required: true
          schema:
            type: string
          description: School Id identifier
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        500:
          description: Internal server error
  /admin/students/school/{schoolId}:
    get:
      tags:
        - Admin
      summary: Get School
      description: "GET /students/school/:schoolId"
      operationId: get_admin_students_school__schoolId
      parameters:
        -
          name: schoolId
          in: path
          required: true
          schema:
            type: string
          description: SchoolId identifier
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        500:
          description: Internal server error
...
  /public/reviews/driver/{driverId}:
    get:
      tags:
        - Public
      summary: Get Driver
      description: "GET /reviews/driver/:driverId"
      operationId: get_public_reviews_driver__driverId
      parameters:
        -
          name: driverId
          in: path
          required: true
          schema:
            type: string
          description: DriverId identifier
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        500:
          description: Internal server error
...
  /shared/schools/{school_id}:
    get:
      tags:
        - Shared
      summary: Get Schools
      description: "GET /schools/:school_id"
      operationId: get_shared_schools__school_id
      parameters:
        -
          name: school_id
          in: path
          required: true
          schema:
            type: string
          description: School Id identifier
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        500:
          description: Internal server error
...
components:
  securitySchemes:
    ParentAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Parent JWT token
    DriverAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Driver JWT token
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT token
...